refactor: features/battle <-> pages/battlePage 순환 의존성 제거
- useBattleChat, useUnreadMessages, usePhaseSkip, useBattleTimer, useResize 훅을
  pages/battlePage/hooks → features/battle/hooks 로 이동
- convertChatMessage 유틸을
  pages/battlePage/utils → features/battle/utils 로 이동
- features/battle/components 내 역참조 import 경로 수정
  (ChatSection, BattleTopBar, BattleTimer, BattleSidebar, TeamChangeModal)
- BattleModals.tsx의 잘못된 상대경로 import 수정
  (DiscussionModal, TeamVoteResultModal → @/features/battle 절대경로
