+++ frontend/src/features/battle/components/modals/TeamChangeModal/PhaseFlowCard.tsx
@@ -2,13 +2,6 @@ import Icon from '@/commons/components/Icon';
 import TimelineMessage from './TimelineMessage';
 import type { PhaseFlowCardProps } from './types';
 
-/**
- * 공수 전환 카드 컴포넌트
- *
- * @description
- * 한 번의 공수 전환(공격 → 수비)을 카드 형태로 표시합니다.
- * 헤더에 공격/수비 팀 정보를 표시하고, 내용은 2컬럼으로 분할하여 렌더링합니다.
- */
 export default function PhaseFlowCard({ attackTeam, defenseTeam, attackMessage, defenseMessage }: PhaseFlowCardProps) {
   const isAttackTeamA = attackTeam === 'A';
 
@@ -25,10 +18,8 @@ export default function PhaseFlowCard({ attackTeam, defenseTeam, attackMessage, 
 
   return (
     <div className="relative bg-[#1a1a2e] rounded-xl border-2 border-[#2d2d3f] overflow-hidden mb-4">
-      {/* 헤더: 공격/수비 팀 정보 */}
       <div className={`px-4 py-2 bg-gradient-to-r ${headerGradient} border-b border-[#2d2d3f]`}>
         <div className="flex items-center justify-center gap-3">
-          {/* 공격 팀 */}
           <div className="flex items-center gap-2">
             <Icon name="zap" className={`w-4 h-4 ${attackIconColor}`} />
             <span className={`${attackTextColor} font-bold text-xs uppercase tracking-wider`}>
@@ -36,10 +27,8 @@ export default function PhaseFlowCard({ attackTeam, defenseTeam, attackMessage, 
             </span>
           </div>
 
-          {/* 화살표 */}
           <Icon name="arrowRight" className="w-4 h-4 text-gray-500" />
 
-          {/* 수비 팀 */}
           <div className="flex items-center gap-2">
             <Icon name="shield" className={`w-4 h-4 ${defenseIconColor}`} />
             <span className={`${defenseTextColor} font-bold text-xs uppercase tracking-wider`}>{defenseTeam} 반론</span>
@@ -47,14 +36,11 @@ export default function PhaseFlowCard({ attackTeam, defenseTeam, attackMessage, 
         </div>
       </div>
 
-      {/* 내용: 공격 | 수비 2컬럼 레이아웃 */}
       <div className="grid grid-cols-1 md:grid-cols-2">
-        {/* 공격 메시지 */}
         <div className={`border-b md:border-b-0 md:border-r border-[#2d2d3f] bg-gradient-to-r ${attackBgGradient}`}>
           <TimelineMessage message={attackMessage} team={attackTeam} type="challenge" />
         </div>
 
-        {/* 수비 메시지 */}
         <div className={`bg-gradient-to-l ${defenseBgGradient}`}>
           <TimelineMessage message={defenseMessage} team={defenseTeam} type="rebuttal" />
         </div>
