+++ frontend/src/features/battle/hooks/useBattleTimer.ts
