+++ frontend/src/features/battle/hooks/useResize.ts
