+++ frontend/src/features/battle/utils/convertChatMessage.ts
